chore: Add bcryptjs, ioredis, and neo4j-driver dependencies
Add the following dependencies to the project:
- bcryptjs@^2.4.3
- ioredis@^5.4.1
- neo4j-driver@^5.20.0

Also update the devDependencies with the latest version of @types/bcryptjs.

Create a new file UnregisteredUser.ts in the src/data/models/bodyData directory, which defines the interface for an unregistered user.

Create a new file Constants.ts in the src/data/neo4j directory, which exports constants related to the Neo4j database.

Create a new file index.ts in the src/data/neo4j directory, which initializes the Neo4j driver using the constants from Constants.ts.

Create a new file users.ts in the src/data/neo4j directory, which contains functions used to interact with the neo4j database.

Create a new file index.ts in the src/data/redis directory, which initializes the Redis client.

Update the index.ts file in the src directory to import UserRoutes and users from the routes/user directory, and redisClient from the data/redis directory.